Moved statement and parameters to valid location
